+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `OxidEsales\EshopCommunity\Core\UtilsView:getTemplateCompileId()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:parseThroughSmarty()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:_smartyDefaultTemplateHandler()`
+  - `OxidEsales\EshopCommunity\Core\UtilsView:getSmartyPluginDirectories()`
+  - `OxidEsales\EshopCommunity\Core\UtilsView:getShopSmartyPluginDirectories()`
   - `oxfunctions:ox_get_template()`
   - `oxfunctions:ox_get_timestamp()`
   - `oxfunctions:ox_get_secure()`